feat : RDCC-1590 code review comments
