+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
     ext {
-        springBootVersion = '2.3.1.RELEASE'
+        springBootVersion = '2.2.6.RELEASE'
     }
     repositories {
         jcenter()
@@ -40,7 +40,7 @@ def versions = [
         reformS2sClient    : '3.0.0',
         serenity           : '2.0.23',
         sonarPitest        : '0.5',
-        springBoot         : '2.3.1.RELEASE',
+        springBoot         : '2.2.6.RELEASE',
         springHystrix      : '2.1.1.RELEASE',
         springfoxSwagger   : '2.9.2',
         camel              : '2.25.1'
