+++ src/integrationTest/java/uk/gov/hmcts/reform/juddata/cameltest/testsupport/IntegrationTestSupport.java
@@ -1,25 +1,8 @@
 package uk.gov.hmcts.reform.juddata.cameltest.testsupport;
 
-import static org.apache.commons.lang3.StringUtils.isBlank;
-
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.sql.Timestamp;
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 import org.springframework.util.ResourceUtils;
-import uk.gov.hmcts.reform.juddata.camel.binder.JudicialOfficeAuthorisation;
 
 public interface IntegrationTestSupport {
-
-    static final String BLANK = "";
 
     static void setSourcePath(String path, String propertyPlaceHolder) throws Exception {
 
@@ -34,54 +17,6 @@ public interface IntegrationTestSupport {
                     + result + "?fileName=" + fileName + "&noop=true");
         } else {
             System.setProperty(propertyPlaceHolder, "file:" + loadFile.replaceFirst("/", ""));
-        }
-    }
-
-    static List<JudicialOfficeAuthorisation> getFileAuthorisationObjectsFromCsv(String inputFilePath)  {
-        List<JudicialOfficeAuthorisation> authorisations = new LinkedList<>();
-        try {
-            File file = ResourceUtils.getFile(inputFilePath);
-            InputStream inputStream = new FileInputStream(file);
-            BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
-            authorisations = bufferedReader.lines().skip(1).map(line -> mapJudicialOfficeAuthorisation(line)).collect(Collectors.toList());
-            bufferedReader.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return authorisations;
-    }
-
-    static JudicialOfficeAuthorisation mapJudicialOfficeAuthorisation(String line) {
-        List<String> columns = Arrays.asList(line.split("\\,", -1));
-        JudicialOfficeAuthorisation judicialOfficeAuthorisation = new JudicialOfficeAuthorisation();
-        judicialOfficeAuthorisation.setElinksId(handleNull(columns.get(0), false));
-        judicialOfficeAuthorisation.setJurisdiction(handleNull(columns.get(1), false));
-        judicialOfficeAuthorisation.setTicketId(isBlank(columns.get(2)) ? null : Long.parseLong(columns.get(2)));
-        judicialOfficeAuthorisation.setStartDate(handleNull(columns.get(3), true));
-        judicialOfficeAuthorisation.setEndDate(handleNull(columns.get(4), true));
-        judicialOfficeAuthorisation.setCreatedDate(handleNull(columns.get(5), true));
-        judicialOfficeAuthorisation.setLastUpdated(handleNull(columns.get(6), true));
-        judicialOfficeAuthorisation.setLowerLevel(handleNull(columns.get(7), false));
-        return judicialOfficeAuthorisation;
-    }
-
-    static String handleNull(String fieldValue, boolean timeStampField) {
-        if (isBlank(fieldValue) && !timeStampField) {
-            return BLANK;
-        } else if (timeStampField && isBlank(fieldValue)) {
-            return null;
-        } else if (timeStampField && !isBlank(fieldValue)) {
-            return Timestamp.valueOf(fieldValue).toString();
-        } else {
-            return fieldValue;
-        }
-    }
-
-    static String handleNull(Timestamp timestamp) {
-        if (Objects.isNull(timestamp)) {
-            return null;
-        } else {
-            return String.valueOf(timestamp);
         }
     }
 }
