+++ src/integrationTest/resources/application-integration.yml
@@ -62,7 +62,6 @@ route:
     processor-class: JudicialUserProfileProcessor
     mapper-class: JudicialUserProfileRowMapper
     csv-binder-object: JudicialUserProfile
-    child-route: judicial-office-appointment,judicial-office-authorisation
 
   judicial-office-appointment:
     id: judicial-office-appointment
