+++ src/main/resources/application-jrd-router-orchestartion.yaml
@@ -20,7 +20,6 @@ route:
     processor-class: JudicialUserProfileProcessor
     mapper-class: JudicialUserProfileRowMapper
     csv-binder-object: JudicialUserProfile
-    child-route: judicial-office-appointment,jdicial-office-authorisation
 
   judicial-office-appointment:
     id: judicial-office-appointment
